devx: split up automod code
