+++ events/userUpdate.ts
@@ -1,4 +1,4 @@
-import { changeNickname } from "../common/automod.js";
+import changeNickname from "../common/nicknames.js";
 import CONSTANTS from "../common/CONSTANTS.js";
 import log from "../common/logging.js";
